Make num_ranks int and some members private

+++ relearn/source/Random.cpp
@@ -11,6 +11,6 @@
 #include "Random.h"
 
 namespace randomNumberSeeds {
-int64_t partition;
-int64_t octree;
+unsigned int partition;
+unsigned int octree;
 } // namespace randomNumberSeeds
